requirements.txt
APScheduler==3.8.1
discord.py==1.7.3
gspread==4.0.1
oauth2client==4.1.3
pandas==1.3.2
python-dotenv==0.19.2
